AdminConsumer's typeCounter has changed from integer to long, following this in the c'tor here (part of JIRA:COMP-8995 changes)
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@186423 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acsGUIs/eventGUI/src/alma/acs/eventbrowser/views/EventData.java
@@ -40,7 +40,7 @@ public class EventData extends AbstractEventData {
 	private final Any eventAny;
 
 	
-	public EventData(long time, String srcObj, long count, String type, Integer typeCount, String chanName, Any any) {
+	public EventData(long time, String srcObj, long count, String type, Long typeCount, String chanName, Any any) {
 		timestamp = time;
 		sourceObject = srcObj;
 		channelEventCount = count;
